+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - 增加数据访问层实现。
   - com.dwarfeng.subgrade.impl.dao.HibernateHqlPresetLookupDao。
+  - com.dwarfeng.subgrade.impl.dao.HibernateAccelerableHqlPresetLookupDao。
 
 - 规范代码的注解、注释、toString 方法。
   - com.dwarfeng.subgrade.impl.dao.HibernateAccelerablePresetLookupDao。